feat(universes/universes_pools): add update method for fetching latest deployment status
update all method can be called on page refresh to update the statuses of all universes and pools
loaded. It is not automatically included in the getAll handler since this would add excessive
latency for GET requests, SQLAdmin and CloudBuild requests to what would otherwise be a database
lookup. i.e. call updateAll and then refresh, as needed.